feat(users): delete now checks levels before allowing delete

+++ sail/Models/User.php
@@ -886,13 +886,91 @@ class User extends Model
     {
         $this->hasPermissions();
 
+        // Do not delete anonymous user
         if ((string)$id === (string)self::anonymousUser()->_id) {
             throw new DatabaseException('9004: Anonymous user cannot be deleted.');
+        }
+
+        // Are we allowed to delete the user?
+        $currentLevel = Role::getHighestLevel(self::$currentUser->roles);
+        $user = self::get($id);
+
+        if (!$user) {
+            return false;
+        }
+
+        $userRole = Role::getHighestLevel($user->roles);
+
+        if ($userRole > $currentLevel) {
+            return false;
         }
 
         $id = $this->ensureObjectId($id);
         $this->deleteById($id);
         Event::dispatch(self::EVENT_DELETE, (string)$id);
+        return true;
+    }
+
+    /**
+     *
+     * Delete a list of users
+     *
+     * @param  array|Collection  $ids
+     * @return bool
+     * @throws ACLException
+     * @throws DatabaseException
+     * @throws PermissionException
+     *
+     */
+    public function removeByIdList(array|Collection $ids): bool
+    {
+        $this->hasPermissions();
+
+        if (is_object($ids)) {
+            $ids = $ids->unwrap();
+        }
+
+        $anonymous = (string)self::anonymousUser()->_id;
+
+        // Remove anonymous from array of ids
+        if (in_array($anonymous, $ids, true)) {
+            array_filter($ids, function ($id) use ($anonymous)
+            {
+                return ($id !== $anonymous);
+            });
+        }
+
+        // Are we allowed to delete the user?
+        $currentLevel = Role::getHighestLevel(self::$currentUser->roles);
+
+        // Remove all users that are not allowed (higher level)
+        foreach ($ids as $num => $id) {
+            $user = self::get($id);
+
+            if (!$user) {
+                unset($ids[$num]);
+                continue;
+            }
+
+            $userRole = Role::getHighestLevel($user->roles);
+
+            if ($userRole > $currentLevel) {
+                unset($ids[$num]);
+            }
+        }
+
+        $ids = array_values($ids);
+
+        // Nothing to delete
+        if (count($ids) === 0) {
+            return false;
+        }
+
+        $list = $ids;
+        $ids = $this->ensureObjectIds($ids)->unwrap();
+        $this->deleteMany(['_id' => ['$in' => $ids]]);
+        
+        Event::dispatch(self::EVENT_DELETE, $list);
         return true;
     }
